+++ DatadogSessionReplay/Sources/Recorder/ViewTreeSnapshotProducer/ViewTreeSnapshot/NodeRecorders/SwiftUI/SwiftUIWireframesBuilder.swift
@@ -24,6 +24,8 @@ internal struct SwiftUIWireframesBuilder: NodeWireframesBuilder {
     let renderer: DisplayList.ViewUpdater
     /// Image renderer for drawings.
     let imageRenderer: ImageRenderer
+    /// Resource builder for captured shapes.
+    let shapeResourceBuilder: ShapeResourceBuilder
     /// Text obfuscator for masking text.
     let textObfuscator: TextObfuscating
     /// Flag that determines if font should be scaled.
@@ -101,20 +103,27 @@ internal struct SwiftUIWireframesBuilder: NodeWireframesBuilder {
         let id: Int64 = .positiveRandom(using: &generator)
 
         switch content.value {
-        case let .shape(_, paint, _):
-            return paint.paint.map { paint in
-                context.builder.createShapeWireframe(
+        case let .shape(path, color, fillStyle):
+            // We treat shapes as bundled images
+            if case .maskAll = self.imagePrivacyLevel {
+                return context.builder.createPlaceholderWireframe(
                     id: id,
                     frame: context.convert(frame: item.frame),
                     clip: context.clip,
-                    backgroundColor: CGColor(
-                        red: CGFloat(paint.linearRed),
-                        green: CGFloat(paint.linearGreen),
-                        blue: CGFloat(paint.linearBlue),
-                        alpha: CGFloat(paint.opacity)
-                    ),
-                    cornerRadius: viewInfo?.cornerRadius,
-                    opacity: CGFloat(paint.opacity)
+                    label: "Image"
+                )
+            } else {
+                let shapeResource = shapeResourceBuilder.shapeResource(
+                    for: path,
+                    color: color,
+                    fillStyle: fillStyle,
+                    size: item.frame.size
+                )
+                return context.builder.createImageWireframe(
+                    id: id,
+                    resource: shapeResource,
+                    frame: context.convert(frame: item.frame),
+                    clip: context.clip
                 )
             }
         case let .text(view, _):